Implement Reader and clean up code
